Move InputPair to the receive module
