+++ payjoin/src/lib.rs
@@ -48,6 +48,5 @@ mod uri;
 
 #[cfg(feature = "base64")]
 pub use bitcoin::base64;
-pub use psbt::InputPair;
 pub use uri::{PjParseError, PjUri, PjUriBuilder, Uri, UriExt};
 pub use url::{ParseError, Url};
