+++ payjoin/src/receive/mod.rs
@@ -29,7 +29,7 @@ use std::cmp::{max, min};
 use bitcoin::base64::prelude::BASE64_STANDARD;
 use bitcoin::base64::Engine;
 use bitcoin::psbt::Psbt;
-use bitcoin::{Amount, FeeRate, OutPoint, Script, TxIn, TxOut, Weight};
+use bitcoin::{psbt, AddressType, Amount, FeeRate, OutPoint, Script, TxIn, TxOut, Weight};
 
 mod error;
 mod optional_parameters;
@@ -47,10 +47,48 @@ use error::{
 };
 use optional_parameters::Params;
 
-use crate::psbt::{InputPair, PsbtExt};
+use crate::psbt::{InternalInputPair, PsbtExt, PsbtInputError};
 
 pub trait Headers {
     fn get_header(&self, key: &str) -> Option<&str>;
+}
+
+/// Helper to construct a pair of (txin, psbtin) with some built-in validation
+/// Use with [`InputPair::new`] to contribute receiver inputs.
+#[derive(Clone, Debug)]
+pub struct InputPair {
+    pub(crate) txin: TxIn,
+    pub(crate) psbtin: psbt::Input,
+}
+
+impl InputPair {
+    pub fn new(txin: TxIn, psbtin: psbt::Input) -> Result<Self, PsbtInputError> {
+        let input_pair = Self { txin, psbtin };
+        let raw = InternalInputPair::from(&input_pair);
+        raw.validate_utxo(true)?;
+        let address_type = raw.address_type()?;
+        if address_type == AddressType::P2sh && input_pair.psbtin.redeem_script.is_none() {
+            return Err(PsbtInputError::NoRedeemScript);
+        }
+        Ok(input_pair)
+    }
+
+    pub(crate) fn address_type(&self) -> AddressType {
+        InternalInputPair::from(self)
+            .address_type()
+            .expect("address type should have been validated in InputPair::new")
+    }
+
+    pub(crate) fn previous_txout(&self) -> TxOut {
+        InternalInputPair::from(self)
+            .previous_txout()
+            .expect("UTXO information should have been validated in InputPair::new")
+            .clone()
+    }
+}
+
+impl<'a> From<&'a InputPair> for InternalInputPair<'a> {
+    fn from(pair: &'a InputPair) -> Self { Self { psbtin: &pair.psbtin, txin: &pair.txin } }
 }
 
 /// The sender's original PSBT and optional parameters
